Increasing the number of queues has shown a reduction in pod creation
time as well as getting rid of "Timed out waiting for annotations"
errors. In some basic scale tests, assuming heavy load on the cluster
(300ms for each logical port add) and adding 5k pods:

15 pod handlers  - 117s
150 pod handlers - 16s

Under downstream real scale testing, we see a drop from 30.5 seconds
average pod ready time, to 20.8 seconds. The CPU remained about the same
throughout the run, but there is an increased memory cost. I estimate
around 100MB increase in ovnkube-master with this patch during scale
testing of 20 nodes with 5k pods created.

Signed-off-by: Tim Rozet <trozet@redhat.com>