@girishmg 2021-07-27
/hold @winsopc is running scale tests against our 700 nodes cluster and will get back with results from his tests