@girishmg 2021-07-28
@trozet Winson came back with these numbers. He is on PTO till Aug 3rd, so sharing it on his behalf.

Baseline (Time taken to bring up 3000 Pods)
```
Tue Jul 27 10:17:56 PDT 2021
303
Tue Jul 27 10:18:57 PDT 2021
1117
Tue Jul 27 10:19:58 PDT 2021
2204
Tue Jul 27 10:21:00 PDT 2021
2908
Tue Jul 27 10:22:02 PDT 2021
2991
Tue Jul 27 10:23:04 PDT 2021
3000
```

With this PR (Time taken to bring up 3000 Pods)
```
Tue Jul 27 20:21:59 PDT 2021
286
Tue Jul 27 20:23:00 PDT 2021
1555
Tue Jul 27 20:24:01 PDT 2021
2770
Tue Jul 27 20:25:03 PDT 2021
3000
```

So, it is significantly faster...