@trozet 2021-07-30
After some more testing/investigation, found that although this gave us some better scale, it was not the right solution. The enqueueEvents are not fairly distributed to each queue, so after changing that, we do not need as many queues. In fact, more queues only increases contention across handler threads, and results in higher variation of pod ready consistency. Going to close this and open another PR with a batch of changes that fix these issues.